Move map_displacements_point and map_loads_point to KynemaFMBBase
