+++ src/aero/fmb/CMakeLists.txt
@@ -1,3 +1,4 @@
 target_sources(${kynema_ugf_lib_name} PRIVATE
+  ${CMAKE_CURRENT_SOURCE_DIR}/KynemaFMBBase.C
   ${CMAKE_CURRENT_SOURCE_DIR}/KynemaFMBSixDof.C  
 )
